\name{kdfe-package}
\alias{kdfe}
\docType{package}
\title{
  kdfe - Kernel density functionals estimation package
}
\description{
  Includes the following functionality:
}
\details{
  More detailed information and/or examples are given per function as needed.
}
\author{ 
  Su Chen, \email{schen4@memphis.edu}
  Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\references{
  \itemize{
    \item TODO
  }
}
\keyword{package}

\seealso{
  \itemize{
    \item \code{\link{rot.kdfe}} - Description
    \item \code{\link{dpi.kdfe}} - Description
  }
}
